fix: make lift/drag coefficient returns version-robust for mypy
The no-any-return ignores on _lift_coefficient/_drag_coefficient depended on
whether the installed numba typed the njit _dynamic_pressure as returning float
or Any, so warn_unused_ignores failed under CI's numba on Python 3.11 while
passing locally on 3.12. Wrap the returns in float() and drop the ignores so the
type is explicit on every version.

+++ src/rocket_tools/aerodynamics/fundamentals.py
@@ -40,7 +40,7 @@ def _lift_coefficient(lift: float, rho: float, v: float, s: float) -> float:
     q = _dynamic_pressure(rho, v)
     if q <= 0.0 or s <= 0.0:
         raise ValueError("Dynamic pressure and area must be > 0")
-    return lift / (q * s)  # type: ignore[no-any-return]
+    return float(lift / (q * s))
 
 
 @njit(cache=True)
@@ -48,7 +48,7 @@ def _drag_coefficient(drag: float, rho: float, v: float, s: float) -> float:
     q = _dynamic_pressure(rho, v)
     if q <= 0.0 or s <= 0.0:
         raise ValueError("Dynamic pressure and area must be > 0")
-    return drag / (q * s)  # type: ignore[no-any-return]
+    return float(drag / (q * s))
 
 
 @njit(cache=True)
